Add copyright headers and move to separate file
