+++ sdk/core/azure_core/src/process/mod.rs
@@ -1,3 +1,6 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
 use async_trait::async_trait;
 use std::io::Error;
 use std::{ffi::OsStr, process::Output};
@@ -12,23 +15,7 @@ pub use thread::ThreadRunner;
 pub use tokio::TokioRunner;
 
 #[cfg(feature = "tokio_process")]
-mod tokio {
-    use super::*;
-
-    /// Implement [`ProcessRunner`] via [`::tokio::process`]
-    #[derive(Debug)]
-    pub struct TokioRunner;
-
-    #[cfg_attr(target_arch = "wasm32", async_trait(?Send))]
-    #[cfg_attr(not(target_arch = "wasm32"), async_trait)]
-    impl ProcessRunner for TokioRunner {
-        async fn run_command(&self, program: &OsStr, args: &[&OsStr]) -> Result<Output, Error> {
-            let mut cmd = ::tokio::process::Command::new(program);
-            cmd.args(args);
-            cmd.output().await
-        }
-    }
-}
+mod tokio;
 
 /// Obtains a new process runner
 pub fn new_process_runner() -> Box<dyn ProcessRunner> {
